requirements.txt
python-telegram-bot==20.4
Pillow
numpy
ultralytics
asyncpg
matplotlib